+++ code/ryzom/client/src/interface_v3/ctrl_text_button.h
@@ -47,7 +47,7 @@ public:
 
 	virtual void updateCoords();
 
-	virtual uint32 getMemory() { return sizeof(*this)+_Id.size(); }
+	virtual uint32 getMemory() { return (uint32)(sizeof(*this)+_Id.size()); }
 
 	// Display part
 	virtual void draw();
